refactor: only use bodyParser for the POST route
While express middleware can be expressed through `app.use()`, this can
get a little costly to run through the whole stack of middleware for
every request. Therefore it is more performant to use the middleware
only for the routes that need it.

This change moves the bodyParser middleware from the `app.use` central
stack, and into the `app.post` routes, so each route that needs to parse
post bodies can, but the `app.get` route does not need to.