fix: mount whole auth app behind route
Mounting the whole app bheind its own route (/login) prevents a bug
where the auth's bodyParser gets run before the editLinks version of
that middleware. The middleware has built in functionality to not parse
twice - which is fine if it is used multiple times with the same
configuration, but in this case it is used with different
configurations.

Essentially this change means that the auth's bodyParser middleware only
gets run for routes that match `/login`, and not for other routes.