fix: reorganise db ID generation logic
Before this change, the first thing the ID generation logic would do is
make a request to see if the ID was taken - except the ID was "", so it
was never taken.

This reorganises the logic so the first part of the loop is to generate
an ID, then check if that _generated ID is taken_. This logic means
there is an actual ID to look up first.